+++ app/Controllers/Admin/AdminHistoryController.php
@@ -1,0 +1,27 @@
+<?php 
+
+
+namespace OrangDalam\PeminjamanRuangan\Controllers\Admin;
+
+use OrangDalam\PeminjamanRuangan\Core\Controller;
+
+class AdminHistoryController extends Controller {
+
+    public function __construct() {
+        $middlewareInstance = $this->middleware('AuthMiddleware');
+        $middlewareInstance->handleAdmin();
+    }
+
+    public function showHistory(): void
+    {
+        if (!($this->loginCheck())) {
+            header('Location: /login');
+            exit();
+        } 
+        
+
+        $this->view('admin/history');
+    }
+
+
+}
