--- app/Controllers/DashboardController.php
@@ -1,115 +1,0 @@
-<?php
-
-namespace OrangDalam\PeminjamanRuangan\Controllers;
-
-use OrangDalam\PeminjamanRuangan\Core\Controller;
-use OrangDalam\PeminjamanRuangan\Models\Jadwal;
-use OrangDalam\PeminjamanRuangan\Models\Peminjaman;
-use OrangDalam\PeminjamanRuangan\Models\Ruang;
-
-class DashboardController extends Controller
-{
-    private Jadwal $jadwal;
-    private Ruang $ruang;
-
-    public function __construct()
-    {
-        $middlewareInstance = $this->middleware('AuthMiddleware');
-        $middlewareInstance->handleUser();
-        $this->jadwal = new Jadwal();
-        $this->ruang = new Ruang();
-    }
-
- 
-
-  
-
-    public function showPinjamPage(): void
-    {
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/pinjam');
-    }
-
-    public function showInboxPage(): void
-    {
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/inbox');
-    }
-
-    public function ShowHistoryPage(): void
-    {
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/history');
-    }
-
-    public function ShowRequestPage(): void
-    {
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/konfirmasiRuangan');
-    }
-
-    public function showJadwalByRuangan($kodeRuang, $namaHari)
-    {
-        return $this->jadwal->getJadwalByRuangDanHari($kodeRuang, $namaHari);
-    }
-
-
-    public function denah($lantai, $bagian, $posisi)
-    {
-        $data = array();
-        foreach ($this->ruang->show($lantai, $bagian, $posisi) as $item) {
-            $data[$item['kode']] = 'disable';
-        }
-        return $data;
-    }
-
-    public function showDenah()
-    {
-        $this->view('user/dashboard');
-    }
-
-    public function showRequestProfile(): void
-    {
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/profile');
-    }
-
-    // detail ruangan, tinggal href 
-    public function showRequesDetailRuangan(): void
-    {
-
-        $kode = $_GET['kode'];
-        if (!isset($kode)) {
-            header('Location: /dashboard');
-            exit;
-        }
-
-        if (!($this->loginCheck())) {
-            header('Location: /login');
-            exit();
-        }
-
-        $this->view('user/detailRuangan');
-    }
-
-}
