+++ app/Controllers/User/InboxController.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace OrangDalam\PeminjamanRuangan\Controllers\User;
+
+use OrangDalam\PeminjamanRuangan\Core\Controller;
+use OrangDalam\PeminjamanRuangan\Models\Jadwal;
+use OrangDalam\PeminjamanRuangan\Models\Peminjaman;
+use OrangDalam\PeminjamanRuangan\Models\Ruang;
+
+class InboxController extends Controller
+{
+    private Jadwal $jadwal;
+    private Ruang $ruang;
+
+    public function __construct()
+    {
+        $middlewareInstance = $this->middleware('AuthMiddleware');
+        $middlewareInstance->handleUser();
+        $this->jadwal = new Jadwal();
+        $this->ruang = new Ruang();
+    }
+
+    public function showInboxPage(): void
+    {
+        if (!($this->loginCheck())) {
+            header('Location: /login');
+            exit();
+        }
+
+        $this->view('user/inbox');
+    }
+
+}
